Move Maven dependencies management to its own package to have clear separation between Maven and Gradle
