+++ src/main/java/tech/jhipster/lite/module/infrastructure/secondary/javadependency/maven/InvalidPomException.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.module.infrastructure.secondary.javadependency;
+package tech.jhipster.lite.module.infrastructure.secondary.javadependency.maven;
 
 import tech.jhipster.lite.error.domain.GeneratorException;
 
@@ -6,7 +6,7 @@ class InvalidPomException extends GeneratorException {
 
   public InvalidPomException() {
     super(
-      badRequest(JavaDependencyErrorKey.INVALID_POM)
+      badRequest(MavenDependencyErrorKey.INVALID_POM)
         .message("Your pom.xml file is invalid, you have to define, at least, an artifact id for your project")
     );
   }
