+++ src/main/java/tech/jhipster/lite/module/infrastructure/secondary/javadependency/maven/MalformedAdditionalInformationException.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.module.infrastructure.secondary.javadependency;
+package tech.jhipster.lite.module.infrastructure.secondary.javadependency.maven;
 
 import tech.jhipster.lite.error.domain.GeneratorException;
 
@@ -6,8 +6,8 @@ class MalformedAdditionalInformationException extends GeneratorException {
 
   public MalformedAdditionalInformationException(Throwable cause) {
     super(
-      internalServerError(JavaDependencyErrorKey.MALFORMED_ADDITIONAL_INFORMATION)
-        .message("Malfomed XML additional elements for plugin")
+      internalServerError(MavenDependencyErrorKey.MALFORMED_ADDITIONAL_INFORMATION)
+        .message("Malformed XML additional elements for plugin")
         .cause(cause)
     );
   }
