+++ src/test/java/tech/jhipster/lite/module/infrastructure/secondary/javadependency/maven/InvalidPomExceptionTest.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.module.infrastructure.secondary.javadependency;
+package tech.jhipster.lite.module.infrastructure.secondary.javadependency.maven;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -16,6 +16,6 @@ class InvalidPomExceptionTest {
     assertThat(exception.getMessage())
       .isEqualTo("Your pom.xml file is invalid, you have to define, at least, an artifact id for your project");
     assertThat(exception.status()).isEqualTo(ErrorStatus.BAD_REQUEST);
-    assertThat(exception.key()).isEqualTo(JavaDependencyErrorKey.INVALID_POM);
+    assertThat(exception.key()).isEqualTo(MavenDependencyErrorKey.INVALID_POM);
   }
 }
