+++ src/test/java/tech/jhipster/lite/module/infrastructure/secondary/javadependency/maven/MavenTypeTest.java
@@ -1,9 +1,10 @@
-package tech.jhipster.lite.module.infrastructure.secondary.javadependency;
+package tech.jhipster.lite.module.infrastructure.secondary.javadependency.maven;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 import tech.jhipster.lite.UnitTest;
+import tech.jhipster.lite.module.infrastructure.secondary.javadependency.maven.MavenType;
 
 @UnitTest
 class MavenTypeTest {
